pyproject.toml
[project]
name = "hive-to-iceberg"
version = "0.1.0"
description = "Migrate Hive tables to Iceberg format (S3 Tables / Glue / MinIO)"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pyspark>=3.5,<4",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.scripts]
hive-to-iceberg = "hive_to_iceberg.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hive_to_iceberg"]
